% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snapshot_duckdb.R
\name{create_snapshot_duckdb}
\alias{create_snapshot_duckdb}
\title{Load snapshot files into an in memory duckdb database}
\usage{
create_snapshot_duckdb()
}
\value{
An in memory duckdb database.
}
\description{
Load snapshot files into an in memory duckdb database
}
\examples{
create_snapshot_duckdb()
}
