% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_visit_numbers.R
\name{fix_visit_numbers}
\alias{fix_visit_numbers}
\title{Fix visit number column}
\usage{
fix_visit_numbers(visit_col, how = "both")
}
\arguments{
\item{visit_col}{A visit number vector containing the default values of visit from the snapshot tables e.g. v1, v2, ...}

\item{how}{A string indicating how to fix the visit vector
Options are:
factor - which will create an ordered factor
zero_pad - left pad the visit number with zeros to ensure consistent length (e.g. v1 becomes v01)
both - Apply both factor and zero_pad options.}
}
\value{
A vector containing the transformed visit number vector.
}
\description{
Fix visit number column
}
\examples{
visit_number_col <- c("t1", "v1", "v2", "v3")
fix_visit_numbers(visit_number_col, how = "factor")
}
