% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_site.R
\name{get_current_site}
\alias{get_current_site}
\title{Get each participants current follow-up site}
\usage{
get_current_site()
}
\value{
A tibble with structured participant ID and
}
\description{
Get each participants current follow-up site
}
\examples{
get_current_site()
}
