% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_visit.R
\name{get_latest_visit}
\alias{get_latest_visit}
\title{Get information about each ENDIA participant's latest visit}
\usage{
get_latest_visit()
}
\value{
A tibble containing the ID, latest visit number, date and url, and days since the latest visit
}
\description{
Get information about each ENDIA participant's latest visit
}
\examples{
get_latest_visit()
}
