% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participant_ids.R
\name{get_participant_ids}
\alias{get_participant_ids}
\title{Get person IDs for each structured participant ID}
\usage{
get_participant_ids(
  who = c("gestational_mother", "biological_mother", "infant", "father")
)
}
\arguments{
\item{who}{a character vector: one or more of gestational_mother, biological_mother, infant, father. All IDs are returned by default.}
}
\value{
A tibble with each structured participant ID and the requested person IDs.
}
\description{
Get person IDs for each structured participant ID
}
\examples{
get_participant_ids(who = c("gestational_mother", "biological_mother"))
}
