% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_priority_variable_table.R
\name{get_priority_variables_table}
\alias{get_priority_variables_table}
\title{Load priority variables table}
\usage{
get_priority_variables_table(table_name, participant_subset = "Overall")
}
\arguments{
\item{table_name}{Name of the priority variables table desired}

\item{participant_subset}{Name of the participant subset to return results for (defaults to Overall)}
}
\value{
A tibble containing the dataset passed to the table_name argument.
}
\description{
Load priority variables table
}
\examples{
get_priority_variables_table("infant_priority_variables")
get_priority_variables_table("maternal_samples", participant_subset = "pp017")
}
