% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_processed_data_table.R
\name{get_processed_data_table}
\alias{get_processed_data_table}
\title{Load processed data}
\usage{
get_processed_data_table(table_name, cols, include_table_name = FALSE)
}
\arguments{
\item{table_name}{Name of the processed data table desired.}

\item{cols}{An optional character vector of columns to be included in the tibble.}

\item{include_table_name}{Logical: if TRUE, include the table name as a column in the output tibble.}
}
\value{
A tibble containing the dataset passed to the table_name argument.
}
\description{
Load processed data
}
\examples{
get_processed_data_table("mother_ids")
}
