% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snapshot_list.R
\name{get_snapshot_list}
\alias{get_snapshot_list}
\title{Get a list of snapshot tables}
\usage{
get_snapshot_list(pattern = NULL, cols, include_table_name = FALSE)
}
\arguments{
\item{pattern}{Optional: A REGEX string to match only specified tables.}

\item{cols}{A character vector of columns to include in each table of the output list.}

\item{include_table_name}{Logical: if True, include the table name as a column in the output for each table in the list (default is FALSE).}
}
\value{
A list of snapshot tables.
}
\description{
Get a list of snapshot tables
}
\examples{
get_snapshot_list(pattern = "b[0-9]{1,2}")
}
