% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snapshot_table.R
\name{get_snapshot_table}
\alias{get_snapshot_table}
\title{Load Snapshot Data}
\usage{
get_snapshot_table(table_name, cols, include_table_name = FALSE)
}
\arguments{
\item{table_name}{A string indicating the name of the csv table to be read in.}

\item{cols}{An optional character vector of columns to be included in the tibble.}

\item{include_table_name}{Logical: if TRUE, include the table name as a column in the output tibble.}
}
\value{
A tibble containing the dataset passed to the table_name argument.
}
\description{
Load Snapshot Data
}
\examples{
get_snapshot_table("b1s")
get_snapshot_table("mothers")
}
