% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_visit_urls.R
\name{get_visit_urls}
\alias{get_visit_urls}
\title{Get the ENDIA frontend registry URL for specified visits}
\usage{
get_visit_urls(visits = "all")
}
\arguments{
\item{visits}{One of "all" (default), "pregnancy", "birth" or "postnatal".}
}
\value{
A tibble containing the structured participant ID, the visit number and the URL of the visit.
}
\description{
Get the ENDIA frontend registry URL for specified visits
}
\examples{
get_visit_urls()
get_visit_urls(visits = "pregnancy")
get_visit_urls(visits = "birth")
get_visit_urls(visits = "postnatal")
}
