% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_withdrawn_participants.R
\name{get_withdrawn_participants}
\alias{get_withdrawn_participants}
\title{Get withdrawn patients, either T1D diagnoses or explicit withdrawals.}
\usage{
get_withdrawn_participants()
}
\value{
A tibble of withdrawn ENDIA participants.
}
\description{
Get withdrawn patients, either T1D diagnoses or explicit withdrawals.
}
