% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List File Names in ENDIA Data Locations}
\usage{
list_files(location = c("snapshot", "processed", "priority_variables"))
}
\arguments{
\item{location}{A string, either snapshot, processed or priority_variables, which is the location to search for file names}
}
\value{
A character vector of file names in the specified location
}
\description{
List File Names in ENDIA Data Locations
}
\examples{
list_files(location = "snapshot")
}
