% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_request.R
\name{process_data_request}
\alias{process_data_request}
\title{Process A Completed Data Request Form}
\usage{
process_data_request(data_request_filepath)
}
\arguments{
\item{data_request_filepath}{The filepath of a completed data request form}
}
\value{
A list of extracted priority variable spreadsheets
}
\description{
Process A Completed Data Request Form
}
