% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1_render_continuous.R
\name{table_1_render_continuous}
\alias{table_1_render_continuous}
\title{Render Continuous Variable in Table 1 Package}
\usage{
table_1_render_continuous(x)
}
\arguments{
\item{x}{variable to be rendered}
}
\value{
Use in Table 1 as an argument to correctly format continous variables
}
\description{
Render Continuous Variable in Table 1 Package
}
