% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df_list_to_excel.R
\name{write_df_list_to_excel}
\alias{write_df_list_to_excel}
\title{Write a List of DataFrames to Excel}
\usage{
write_df_list_to_excel(df_list, output_path)
}
\arguments{
\item{df_list}{A named list of dataframes to write out to an xlsx file}

\item{output_path}{The desired path of the excel workbook}
}
\value{
Nothing
}
\description{
Write a List of DataFrames to Excel
}
