% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snapshot_table.R
\name{get_snapshot_table}
\alias{get_snapshot_table}
\title{Load Snapshot Data}
\usage{
get_snapshot_table(file_name)
}
\arguments{
\item{file_name}{A string indicating the name of the csv table to be read in.}
}
\value{
A tibble containing the dataset passed to the file_name argument.
}
\description{
Load Snapshot Data
}
\examples{
get_snapshot_table("b1s")
get_snapshot_table("mothers")
}
